Finish refactor Index.js
Fix issue where card elements added in incorrect order.
This was due to the dfact I was calling aysnc functions without
waiting for them to finish executing.
Add await keyword to the calls for aysnc functions.
Crud Service was being fed cardContainer as param for disaster id
in MySQL query, so fixed that as well.